+++ resources/views/components/navbar-link.blade.php
@@ -1,0 +1,5 @@
+@props([
+    'href' => '#',
+    'active' => false,
+])
+<a href="{{ $href }}" {{ $attributes->merge(['class' => 'vx-navbar-link' . ($active ? ' is-active' : '')]) }}>{{ $slot }}</a>